2004-05-10  Mike Kestner  <mkestner@ximian.com>
	* gtk/Gtk.metadata : mark some Dialog API params as ResponseType
	instead of int.
	[fixes #58240]

svn path=/trunk/gtk-sharp/; revision=27046

+++ sample/GtkDemo/DemoSizeGroup.cs
@@ -149,7 +149,7 @@ namespace GtkDemo
 		private void ResponseCallback (object obj, ResponseArgs args)
         
                {
-		       if (args.ResponseId == 5) {
+		       if (((int) args.ResponseId) == 5) {
 			       window.Hide ();
 			       window.Destroy ();}
 	       }
